♻️  ArticleおよびPost型を新規作成し、api.tsから型定義を移動
