+++ frontend/src/lib/type.ts
@@ -1,0 +1,12 @@
+export type Post = {
+  userId: number;
+  id: number;
+  title: string;
+  body: string;
+};
+
+export type Article = {
+  id: number;
+  title: string;
+  content: string;
+};
